early setup
